uploading fp16

+++ fmengine/utils/hf.py
@@ -1,3 +1,4 @@
+import torch
 from transformers import AutoModel, AutoConfig, AutoTokenizer
 from huggingface_hub import HfApi
 
@@ -6,7 +7,8 @@ def upload_hf(
         repo_id: str,
         revision: str='main'
     ):
-    model = AutoModel.from_pretrained(path_to_hf_model, from_tf=False, use_safetensors=True)
+    model = AutoModel.from_pretrained(path_to_hf_model, from_tf=False, use_safetensors=True, torch_dtype=torch.float16)
+
     model.push_to_hub(repo_id, revision=revision,safe_serialization=True)
     tokenizer = AutoTokenizer.from_pretrained(path_to_hf_model)
     tokenizer.push_to_hub(repo_id, revision=revision)
